requirements.txt
google-cloud-storage==3.7.0
langchain==1.2.3
langchain-community==0.4.1
langchain-text-splitters==1.1.0
docx2txt==0.9
pypdf==6.6.0
vertexai==1.71.1
python-dotenv==1.2.1